🧹 Move documentation code from Command and Flag to Docs
